rename models file
